Fix broken links to prebuilt toolchains on arduino.cc
The problem was that arduino.cc was sending the build tool on a wild
goose chase of redirects.  The original HTTP URL was redirected to an
HTTPS one, which was eventually redirected *back* to an HTTP one. Java
complained that this redirect was "not secure."

Again, this is one of the problems with not having a self-contained
source tree: you break things when you change unrelated things!